requirements.txt
django==4.1.3

django-cors-headers
# pymysql
# mysqlclient==2.1.0

pytest
pytest-django
coverage

uwsgi

channels==4.0.0
daphne==4.0.0

django-composite-foreignkey

websockets
requests

urllib3
aiofiles

pillow